third_party/WebKit/LayoutTests/http/tests/security/suborigins/multiple-suborigins-disallowed-expected.txt

CONSOLE ERROR: Ignoring duplicate Content-Security-Policy directive 'suborigin'.
